+++ packages/backend/app.ts
@@ -1,7 +1,6 @@
 
 import express, { Application } from 'express';
 import cors from 'cors';
-import authGoogleRoutes from './src/routes/authGoogleRoutes';
 import exampleRouts from './src/routes/exampleRouts';
 import userRouts from './src/routes/userRouts';
 import authRouts from './src/routes/authRouts';
@@ -30,7 +29,6 @@ app.use(cookieParser());
 app.use('/api', exampleRouts);
 app.use('/users', userRouts);
 app.use('/auth', authRouts);
-app.use('/auth', authGoogleRoutes);
 
 
 export default app;
