+++ packages/backend/src/routes/authRouts.ts
@@ -1,5 +1,5 @@
 import express from 'express';
-import { generateAndSendCode, login, signup, validateCode, refreshToken, logout, requestSignup, confirmSignup, forgotPassword, resetPassword } from '../controllers/authController';
+import { generateAndSendCode, login, signup, validateCode, refreshToken, logout, requestSignup, confirmSignup, forgotPassword, resetPassword, authWithGoogle } from '../controllers/authController';
 import { authenticateToken } from '../middlewares/authMiddlewares';
 import { getMe } from '../controllers/userController';
 
@@ -8,9 +8,7 @@ const router = express.Router();
 
 router.post('/login', login);
 router.post('/signup', signup);
-//יצירת הקוד ושליחתו במייל
 router.post('/createValidationCode',  generateAndSendCode);
-//בדיקה אם הקוד תקין
 router.post('/validateCode',  validateCode);
 router.post('/refresh', refreshToken);
 router.post('/logout', logout);
@@ -19,6 +17,7 @@ router.post('/signup/request', requestSignup);
 router.post('/signup/confirm', confirmSignup);
 router.post('/forgot-password', forgotPassword);
 router.post('/reset-password', resetPassword);
+router.post('/google-auth', authWithGoogle);
 
 
 export default router;
