+++ packages/backend/src/utils/emailSender.ts
@@ -20,3 +20,24 @@ export async function sendResetEmail(email: string, token: string) {
 
   await transporter.sendMail(mailOptions);
 }
+
+export async function sendVerificationCodeEmail(to: string, text: string): Promise<void> {
+    // Configure your SMTP transporter
+    const transporter = nodemailer.createTransport({
+        host: process.env.SMTP_HOST,
+        port: Number(process.env.SMTP_PORT) || 587,
+        secure: false, 
+        auth: {
+            user: process.env.Email_USER,
+            pass: process.env.Email_PASS,
+        },
+    });
+    console.log('SMTP_HOST:', process.env.SMTP_HOST);
+
+    await transporter.sendMail({
+        from: `"DiversiTech LingoPrep" <${process.env.EMAIL_USER}>`,
+        to: to, 
+        subject: 'קוד אימות',
+        text: text,
+    });
+}
