--- packages/frontend/src/features/auth/components/LoginForm.css
@@ -1,129 +1,0 @@
-.login-form-container {
-  background-color: #e9fdf3; /* רקע ירקרק בהיר */
-  max-width: 400px;
-  margin: 2rem auto;
-  padding: 2rem;
-  border-radius: 16px;
-  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.06);
-  font-family: "Arial", sans-serif;
-  text-align: center;
-}
-
-.login-form-container h2 {
-  font-size: 1.8rem;
-  margin-bottom: 1.5rem;
-  color: #1a1a1a;
-}
-
-.login-form-container input[type="email"],
-.login-form-container input[type="password"] {
-  width: 100%;
-  padding: 0.75rem;
-  margin-bottom: 1rem;
-  border: 1px solid #ccc;
-  border-radius: 12px;
-  font-size: 1rem;
-  background-color: #fff;
-}
-
-.login-form-container input[type="checkbox"] {
-  margin-right: 0.5rem;
-  vertical-align: middle;
-}
-
-.login-form-container label {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-bottom: 1rem;
-  color: #333;
-  font-size: 0.95rem;
-}
-
-.login-form-container button[type="submit"] {
-  width: 100%;
-  padding: 0.75rem;
-  font-size: 1rem;
-  background-color: #00b578;
-  color: white;
-  border: none;
-  border-radius: 12px;
-  cursor: pointer;
-  transition: background-color 0.3s;
-}
-
-.login-form-container button[type="submit"]:hover {
-  background-color: #009e6a;
-}
-
-.login-form-container p {
-  margin-top: 1rem;
-  font-size: 0.95rem;
-}
-
-.google-auth-btn-wrapper {
-  text-align: center;
-  margin-top: 1rem;
-}
-
-.google-auth-btn-wrapper > button,
-.google-auth-btn-wrapper > div {
-  width: 300px;
-  max-width: 100%;
-  margin: 0 auto;
-  display: block;
-}
-
-.google-auth-btn-wrapper iframe {
-  width: 100% !important;
-  min-width: 300px;
-  max-width: 300px;
-  display: block;
-  margin: 0 auto;
-}
-
-/* אימות קוד */
-
-
-
-
-
-.code-verification-input-wrapper {
-  position: relative;
-  display: inline-block;
-  margin-top: 1rem;
-}
-
-.code-verification-input-wrapper input {
-  padding: 0.75rem;
-  padding-right: 2.5rem;
-  font-size: 1.2rem;
-  border: 1px solid #ccc;
-  border-radius: 0.5rem;
-  width: 100%;
-  text-align: center;
-  direction: ltr;
-}
-
-.code-verification-input-wrapper svg {
-  position: absolute;
-  top: 50%;
-  left: 0.75rem;
-  transform: translateY(-50%);
-  cursor: pointer;
-}
-
-.code-verification-message {
-  margin-top: 1rem;
-  color: green;
-}
-
-
-.code-verification-button:hover {
-  background-color: #00a374;
-}
-.code-verification-input-wrapper {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
