--- packages/frontend/src/features/auth/types/formTypes.ts
@@ -1,13 +1,0 @@
-export interface ForgotPasswordForm {
-  email: string;
-}
-
-export interface ResetPasswordForm {
-  token: string;
-  password: string;
-}
-
-export type ResetFormData = {
-  password: string;
-  confirm: string;
-};
