+++ packages/frontend/src/features/auth/types/types.ts
@@ -8,3 +8,17 @@ export interface User {
   createdAt: string;
   isActive: boolean;
 }
+
+export interface ForgotPasswordForm {
+  email: string;
+}
+
+export interface ResetPasswordForm {
+  token: string;
+  password: string;
+}
+
+export type ResetFormData = {
+  password: string;
+  confirm: string;
+};
