+++ packages/frontend/src/index.tsx
@@ -4,7 +4,6 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import { Provider } from 'react-redux';
-import { BrowserRouter } from 'react-router-dom';
 import { store } from './shared/store/store';
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
